[add&update]入力補完機能の追加とリファクタリング
