+++ app/campus-map/components/index.ts
@@ -1,0 +1,3 @@
+export { BuildingSearchForm } from "./BuildingSearchForm";
+export { BuildingSearchResultCard } from "./BuildingSearchResultCard";
+export { GoogleMapIFrame } from "./GoogleMapIFrame";
