+++ app/campus-map/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useBuildingSearch } from "./useBuildingSearch";
